Reformat code.
